Modifying Primary Keys from int to long and impelementing order API

+++ src/main/java/com/rrdlabs/dukaanmanager/repositories/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
-    List<Order> findByCustomerId(int customerId);
+    List<Order> findByCustomerId(Long customerId);
 }
