+++ src/main/java/com/rrdlabs/dukaanmanager/repositories/ProductCategoryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
 
     List<ProductCategory> findByDescriptionContainingIgnoreCase(String description);
