+++ src/main/java/com/rrdlabs/dukaanmanager/repositories/StaffRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StaffRepository extends JpaRepository<Staff, Integer> {
+public interface StaffRepository extends JpaRepository<Staff, Long> {
 
     List<Staff> findByPhoneNo(String phoneNo);
